컴포넌트별로 파일 분리
