+++ src/Button.jsx
@@ -1,0 +1,7 @@
+export default function Button({ children, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      {children}
+    </button>
+  );
+}
